ANDROID: add back test files
Due to a cargo2android bug and various tool
changes, these were not included sometimes.

In two cases (serde_keyvalue_derive and arg_helpers),
we don't have the required dependencies for these
tests setup, b/c the component is a rust_proc_macro.

In the case of rutabaga_gfx, no change is made here
because it does not compile with the in-tree cargo
install.

Bug: 242733327
Test: build only
Change-Id: Ib0a1f80b62286cd5e77f7ff116f839ace34ebc81