ANDROID: Add "rotary" virtio input device
Add support to crosvm for a virtio_input device which presents as a
"rotary" wheel for wearable device virtualization. This is a simple
device that supports only EV_REL + REL_WHEEL.

Because of the way Android's inputflinger classifies devices, the device
must *not* have REL_X, REL_Y or BTN_MOUSE, or else SurfaceFlinger will
display a cursor overlay (mouse pointer) which is not desirable. This is
why the existing "mouse" device cannot be reused.

Inspired by the goldfish emulator feature which is implemented over at:
https://android.googlesource.com/platform/external/qemu/+/emu-master-dev/android-qemu2-glue/emulation/virtio-input-rotary.c

Bug: 284320110
Link: https://crrev.com/c/4571175
Change-Id: I2245e4e65cc0d11d91345e90061d39828cf5a1dc