UPSTREAM: aarch64: Add boot_cpu option
Allow users to start a VM using a specified vCPU thread, this
dramatically improves boot times for system using some form of vCPU to
pCPU pinning.

This is especially useful if a bootloader spends a large amount of time
running single threaded on an asymmetrical system before enabling other
cores.

Bug=b:329911550
Test: tools/presubmit
Change-Id: I81e2ae10769b230b201e98c18cca85daee47148f
Signed-off-by: David Dai <davidai@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5384625
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>