ANDROID: Expose qemu symbols in rutabaga
We now link against the qemu-plugin dependency which exposes the qemu
dependency symbol. (_qemu_module_dummy and _qemu_stamp_xxx)

This will allow us to directly load the rutabaga shared library as
a qemu plugin. This will make sure we do not run into rpath issues when
loading the dependencies.

Bug: 356095570
Change-Id: Ic6de96c802d4598f28652e8e90e2fc334916959a