vhost_user: ignore unknown features from backend
Backport of https://crrev.com/c/5708608.

We need to backport this because some tests are using a crosvm binary
from an old branch with a vhost-user device that advertises new protocol
features. It is OK if the new features are ignored, the vhost-user
protocol was designed to allow for it.

Android on-device does not use crosvm with vhost-user devices, so this
should only affect Cuttlefish and is low risk.

Bug: 360957432
Test: TH
Ignore-AOSP-First: backport from AOSP
Merged-In: Id1bacf21d6b38946f0cf4e0d8c79d60ff946de24
Change-Id: I9fc06f81e96ab4c77bac3a38a499e215e46b6ab8