Revert "balloon: Extend dynamic_mapping tube support to linux"
This reverts commit 22efdcd20ab201e1a18cfb0edc5a02286337eea4.

The upstream balloon code along with the handling of tubes has changed a
lot and this CL is creating messy merge conflicts when trying to merge
the upstream crosvm code into Android.

I don't know of a good way for us to test this functionality, so we will
revert this for now while we work with QCOM to create and upstream
another attempt.

Test: none
Bug: 372302460
Change-Id: Ieae5d5d6a3af8f75ff3b621964a3956666bc9373