UPSTREAM: devices: vhost-user frontend: fix backend_req on Linux
On Linux, we don't need to monitor anything special to get close
notifications; the Token::ReqHandlerRead event will be generated when
the fd closes (with is_hungup = true) and the closure will be handled
inline.

Adding the same backend_req_handler.get_read_notifier() to the wait_ctx
again causes an error, since that fd was already added in a previous
call.

BUG=None
TEST=`crosvm device block` + `crosvm run`

Fixes: 8de3640c01bf ("devices: vhost-user frontend: rewrite worker as non-async")
Change-Id: I1318c5395670476bdc333a2bae1e8c8aba2c658e